+++ backend/src/app.module.ts
@@ -3,8 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm'
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UsersModule } from './users/users.module';
-import { User } from './users/user.entity';
-import { AuthModule } from './auth/auth.module';
+import { User } from './users/entities/user.entity';
 
 @Module({
   imports: [
@@ -14,8 +13,7 @@ import { AuthModule } from './auth/auth.module';
       entities: [User],
       synchronize: true
     }),
-    UsersModule,
-    AuthModule
+    UsersModule
   ],
   controllers: [AppController],
   providers: [AppService],
