--- backend/src/auth/constants.ts
@@ -1,4 +1,0 @@
-//tmp
-export const jwtConstants = {
-    secret: 'secretKey',
-  };
