--- backend/src/dtos/create-user.dto.ts
@@ -1,12 +1,0 @@
-import { IsEmail, IsString } from "class-validator";
-
-export class CreateUserDto {
-    @IsEmail()
-    email: string;
-
-    @IsString()
-    username: string;
-
-    @IsString()
-    password: string;
-}
