--- backend/src/dtos/sign-user-in.dto.ts
@@ -1,9 +1,0 @@
-import { IsString } from "class-validator";
-
-export class SignUserInDto {
-    @IsString()
-    username: string;
-
-    @IsString()
-    password: string;
-}
