+++ backend/src/interceptors/serialize.interceptor.ts
@@ -1,8 +1,12 @@
-import { CallHandler, ExecutionContext,
-    NestInterceptor, UseInterceptors } from "@nestjs/common";
+import { CallHandler, ExecutionContext, NestInterceptor, UseInterceptors } from "@nestjs/common";
 import { Observable } from "rxjs";
 import { map } from "rxjs/operators";
 import { plainToClass } from "class-transformer";
+
+/**
+ *  Create a @Serialize() decorator.
+ *  Takes a DTO that describes how to serialize the output of a route handler.
+ */
 
 interface ClassConstructor {
     new (...args: any[]): {}
