+++ backend/src/request.http
@@ -3,8 +3,8 @@ POST http://localhost:3000/auth/signup
 Content-Type: application/json
 
 {
-    "email": "newmail@mail.com",
-    "username": "name",
+    "email": "lol@mail.com",
+    "username": "k",
     "password": "password"
 }
 
@@ -13,17 +13,20 @@ POST http://localhost:3000/auth/signin
 Content-Type: application/json
 
 {
-    "username": "name",
+    "username": "yayao",
     "password": "password"
 }
 
 ### update
-PATCH http://localhost:3000/users/1
+PATCH http://localhost:3000/auth/1
 Content-Type: application/json
 
 {
-    "username": "newName"
+    "username": "yayao"
 }
+
+### get user by id
+GET http://localhost:3000/users/1
 
 ### get user by name
 GET http://localhost:3000/users?username=newName
