+++ backend/src/users/constants.ts
@@ -1,0 +1,9 @@
+/**
+ *  This key should not be exposed publicly. In a production system,
+ *  it must be protected using appropriate measures such as a secrets
+ *  vault, environment variable, or configuration service.
+ */
+
+export const jwtConstants = {
+    secret: 'secretKey',
+  };
