+++ backend/src/users/dtos/create-user.dto.ts
@@ -1,0 +1,19 @@
+import { IsEmail, IsString, IsNotEmpty } from "class-validator";
+
+/**
+ *  Defines how sign up inputs should be formated.
+ */
+
+export class CreateUserDto {
+    @IsNotEmpty()
+    @IsEmail()
+    email: string;
+
+    @IsNotEmpty()
+    @IsString()
+    username: string;
+
+    @IsNotEmpty()
+    @IsString()
+    password: string;
+}
