+++ backend/src/users/dtos/update-user.dto.ts
@@ -1,14 +1,21 @@
-import { IsEmail, IsString, IsOptional } from "class-validator";
+import { IsEmail, IsString, IsOptional, IsNotEmpty } from "class-validator";
+
+/**
+ *  Defines how update inputs should be formated.
+ */
 
 export class UpdateUserDto {
+    @IsNotEmpty()
     @IsEmail()
     @IsOptional()
     email: string;
 
+    @IsNotEmpty()
     @IsString()
     @IsOptional()
     username: string;
 
+    @IsNotEmpty()
     @IsString()
     @IsOptional()
     password: string;
