+++ backend/src/users/dtos/user.dto.ts
@@ -1,5 +1,9 @@
 import { Expose } from "class-transformer";
 
+/**
+ *  Describes how to serialize a user.
+ */
+
 export class UserDto {
     @Expose()
     id: number;
@@ -9,7 +13,4 @@ export class UserDto {
 
     @Expose()
     username: string;
-
-    @Expose()
-    access_token: string;
 }
