+++ backend/src/users/users.module.ts
@@ -1,14 +1,18 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersService } from './users.service';
-import { User } from './user.entity';
-import { UsersController } from './users.controller';
-import { AuthService } from 'src/auth/auth.service';
+import { UsersService } from './services/users.service';
+import { User } from './entities/user.entity';
+import { UsersController } from './controllers/users.controller';
+import { AuthService } from './services/auth.service';
 
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 
-import { jwtConstants } from '../auth/constants'
+import { jwtConstants } from './constants'
+
+import { JwtStrategy } from './strategies/jwt.strategy'
+
+import { AuthController } from '../users/controllers/auth.controller';
 
 @Module({
     imports: [
@@ -19,8 +23,8 @@ import { jwtConstants } from '../auth/constants'
         }),
         TypeOrmModule.forFeature([User])
     ],
-    providers: [UsersService, AuthService],
-    exports: [UsersService],
-    controllers: [UsersController]
+    providers: [UsersService, AuthService, JwtStrategy],
+    // exports: [UsersService],
+    controllers: [UsersController, AuthController]
 })
 export class UsersModule {}
